# pms_api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.InsertProject:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  model.Message:
    properties:
      message:
        type: string
    type: object
  model.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.ProjectShort:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  model.ProjectsPaged:
    properties:
      items:
        items:
          $ref: '#/definitions/model.ProjectShort'
        type: array
      page_index:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.Task:
    properties:
      author_id:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      executor_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        type: string
      tester_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      position:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: PMS - project management system. It is course work of student team
    for 5 course in MEPhI.
  termsOfService: http://swagger.io/terms/
  title: PMS API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get projects
      parameters:
      - description: Page index of projects
        in: query
        name: pageIndex
        type: string
      - description: Page size of projects
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectsPaged'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Create project
      parameters:
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.InsertProject'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "422":
          description: Bind error
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Create project
      tags:
      - Project
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete project
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Delete project
      tags:
      - Project
    get:
      consumes:
      - application/json
      description: Get project by ID
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get project by ID
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Update project
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.InsertProject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project/Bind error
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Update project
      tags:
      - Project
  /projects/{project_id}/archive:
    put:
      description: Archive the project by id
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Archive the project by id
      tags:
      - Project
  /projects/{project_id}/members:
    get:
      consumes:
      - application/json
      description: Get project members
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get project members
      tags:
      - Project
  /projects/{project_id}/members/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete project member
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found/User not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project/Incorrect id of user
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Delete project member
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Add project member
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found/User not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project/Incorrect id of user
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Add project member
      tags:
      - Project
  /projects/{project_id}/tasks:
    get:
      description: Get project tasks
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get project tasks
      tags:
      - Project
  /projects/{project_id}/unarchive:
    put:
      description: Unarchive the project by id
      parameters:
      - description: Project id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id of project
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Unarchive the project by id
      tags:
      - Project
  /projects/archived:
    get:
      description: Get archived projects
      parameters:
      - description: Page index of projects
        in: query
        name: pageIndex
        type: string
      - description: Page size of projects
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectsPaged'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get archived projects
      tags:
      - Project
...
securityDefinitions:
  Login:
    description: Type "Bearer TOKEN" to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"